requirements1.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arguments==76
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
beartype==0.16.4
black==23.12.1
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpickle==3.0.0
colorama==0.4.6
colormaps==0.4.1
colorspacious==1.1.2
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator==4.4.2
-e git+https://github.com/facebookresearch/detectron2.git@afe9eb920646102f7e6bf0cd2115841cea2aca13#egg=detectron2
diff-gaussian-rasterization @ file:///data/gyy/lsplat/latent-gaussian-rasterization
docker-pycreds==0.4.0
e3nn==0.5.1
easydict==1.11
einops==0.7.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gpustat==1.1.1
gpustats==0.0.1
grpcio==1.60.0
h5py==3.10.0
huggingface-hub==0.21.4
hydra-core==1.3.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
imgviz==1.7.5
importlib_metadata==7.1.0
iopath==0.1.9
ipykernel==6.28.0
ipython==8.20.0
jaxtyping==0.2.25
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.0
lpips==0.1.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mediapy==1.2.0
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nose==1.3.7
numpy==1.26.3
nvidia-ml-py==12.535.133
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
packaging==23.2
pandas==2.2.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.12.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs==4.1.0
plyfile==1.0.3
portalocker==2.8.2
proglog==0.1.10
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==4.23.4
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
pytorch-lightning==2.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
ruff==0.1.13
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scikit-video==1.1.11
scipy==1.11.4
segment-anything @ file:///data/gyy/LangSplat/submodules/segment-anything-langsplat
sentencepiece==0.2.0
sentry-sdk==1.39.2
setproctitle==1.3.3
shapely==2.0.3
simple-knn @ file:///data/gyy/LangSplat/submodules/simple-knn
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.4.0
tifffile==2023.12.9
timm==0.9.16
tomli==2.0.1
torch==1.11.0+cu113
torch-encoding @ git+https://github.com/zhanghang1989/PyTorch-Encoding/@c959dab8312b637fcc7edce83607acb4b0f82645
torchaudio==0.11.0+cu113
torchmetrics==1.3.0
torchvision==0.12.0+cu113
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827254365/work
tqdm==4.66.1
traitlets==5.14.1
transforms3d==0.4.1
typeguard==2.13.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
tzdata==2024.1
urllib3==2.1.0
visdom==0.2.4
wandb==0.16.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
websocket-client==1.7.0
Werkzeug==3.0.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
